fix create question

+++ Application/Usecases/Command/CreateQuestionsCommand.cs
@@ -14,7 +14,7 @@ namespace Application.Usecases.Command
         public string TestSectionID { get; set; }
         public TestFormatType FormatType { get; set; }
         public int NumberOfQuestions { get; set; }
-        public float Score { get; set; }
+        public decimal Score { get; set; }
 
     }
 }
